+++ protoc_plugin/test/freeze_test.dart
@@ -2,8 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:fixnum/fixnum.dart';
 import 'package:test/test.dart';
 
+import '../out/protos/entity.pb.dart';
 import '../out/protos/nested_message.pb.dart';
 
 void main() {
@@ -51,5 +53,70 @@ void main() {
     expect(msg1.isFrozen, true);
     expect(msg2.isFrozen, true);
     expect(msg3.isFrozen, true);
+  });
+
+  test('frozen messages should not be updated by merge methods', () {
+    final top = TopEntity()..freeze();
+
+    expect(() => top.mergeFromBuffer(<int>[]),
+        throwsA(TypeMatcher<UnsupportedError>()));
+
+    expect(() => top.mergeFromJsonMap({}),
+        throwsA(TypeMatcher<UnsupportedError>()));
+
+    expect(() => top.mergeFromMessage(TopEntity()),
+        throwsA(TypeMatcher<UnsupportedError>()));
+
+    expect(() => top.mergeFromProto3Json({}),
+        throwsA(TypeMatcher<UnsupportedError>()));
+  });
+
+  test('nested frozen messages should not be updated by merge methods', () {
+    // Check that recursive calls to merge methods check mutability.
+    // `mergeFromJsonMap` cannot be tested because of #726.
+
+    {
+      final top = TopEntity()..sub = (SubEntity()..freeze());
+
+      top.mergeFromBuffer(<int>[
+        (1 << 3) | 0, // tag = 1, type = varint
+        123, // int64 id = 123
+      ]);
+      expect(top.id, Int64(123));
+
+      expect(
+          () => top.mergeFromBuffer(<int>[
+                (4 << 3) | 2, // tag = 4, type = length delimited
+                2, // length
+                (1 << 3) | 0, // tag = 1, type = varint
+                123, // int64 id = 123
+              ]),
+          throwsA(TypeMatcher<UnsupportedError>()));
+    }
+
+    {
+      final top = TopEntity()..sub = (SubEntity()..freeze());
+
+      top.mergeFromMessage(TopEntity()..id = Int64(123));
+      expect(top.id, Int64(123));
+
+      expect(
+          () => top.mergeFromMessage(
+              TopEntity()..sub = (SubEntity()..id = Int64(123))),
+          throwsA(TypeMatcher<UnsupportedError>()));
+    }
+
+    {
+      final top = TopEntity()..sub = (SubEntity()..freeze());
+
+      top.mergeFromProto3Json({'id': 123});
+      expect(top.id, Int64(123));
+
+      expect(
+          () => top.mergeFromProto3Json({
+                'sub': {'id': 123}
+              }),
+          throwsA(TypeMatcher<UnsupportedError>()));
+    }
   });
 }